requirements.txt
azure-identity
azure-storage-file-share
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-weaviate
weaviate-client
pypdf
docx2txt
